Automatically include versioned service processor for all controllers
